Add additional intelligence and cleanup comments in makesymlinks.sh.

+++ makesymlinks.sh
@@ -1,32 +1,44 @@
 #!/bin/bash
-############################
-# .make.sh
-# This script creates symlinks from the home directory to any desired dotfiles in ~/dotfiles
-############################
 
-########## Variables
+##############################################################################
+# .makesymlinks.sh                                                           #
+#                                                                            #
+# This script creates symlinks from the home directory to any desired        #
+# dot files in ~/dotfiles                                                    #
+##############################################################################
+
+#|    Variables                                                            {{{
+#|============================================================================
 
 dir=~/dotfiles                    # dotfiles directory
 olddir=~/dotfiles_old             # old dotfiles backup directory
-files="irssi mailcap pentadactylrc screenrc tmux.conf vimpagerrc vimrc\
+files="gdbinit irssi mailcap pentadactylrc screenrc tmux.conf vimpagerrc vimrc\
        vim Xdefaults xinitrc xmobarrc xmonad zshenv zshrc zsh"    # list of files/folders to symlink in homedir
+change_counter=0
 
-##########
+# }}} #
 
-# create dotfiles_old in homedir
-echo -n "Creating $olddir for backup of any existing dotfiles in ~ ..."
-mkdir -p $olddir
-echo "done"
+if [ ! -e $olddir ]; then
+    echo -n "Creating $olddir for backup of any existing dotfiles in ~ ..."
+    mkdir -p $olddir
+    echo " done"
+fi
 
-# change to the dotfiles directory
-echo -n "Changing to the $dir directory ..."
 cd $dir
-echo "done"
 
-# move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks from the homedir to any files in the ~/dotfiles directory specified in $files
+# Move any existing dot files in user's home directory to $olddir, then create
+# symlinks from the user's home directory to the files in the ~/dotfiles
+# directory specified in $files
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file $olddir
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/.$file ~/.$file
+    if ! diff -rq ~/.$file $dir/.$file; then
+        if [ -e ~/.$file ]; then
+            echo "Moving existing .$file from ~ to $olddir"
+            mv ~/.$file $olddir
+        fi
+        echo "Creating symlink to $file in home directory."
+        ln -s $dir/.$file ~/.$file
+        change_counter=$((change_counter+1))
+    fi
 done
+
+echo "Changed $change_counter files."
